+++ src/test/java/com/abc/TransactionTest.java
@@ -4,10 +4,12 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
 
+import java.math.BigDecimal;
+
 public class TransactionTest {
     @Test
     public void transaction() {
-        Transaction t = new Transaction(5);
+        Transaction t = new Transaction(BigDecimal.valueOf(5));
         assertTrue(t instanceof Transaction);
     }
 }
